＜変更点＞ ・スポット詳細画面のGridViewが正しく表示できるようになった  （設定されている座標から半径25m以内の写真を表示する） ・マップ画面に処理を追加
